day8-12/pyproject.toml
[project]
name = "shop-service"
version = "1.0.0"
description = "E-commerce Backend — Django Day 8-12"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Django core  (Day 8)
    "django>=5.0.0",
    "django-environ>=0.11.0",
    # Django REST Framework  (Day 9)
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-filter>=24.0",
    "drf-spectacular>=0.27.0",       # Auto API docs
    # Database  (Day 8)
    "psycopg[binary]>=3.1.0",        # PostgreSQL async driver
    # Task Queue  (Day 11)
    "celery>=5.4.0",
    "redis>=5.0.0",
    "django-celery-results>=2.5.0",
    # Utilities
    "pillow>=10.0.0",                 # ImageField
    "structlog>=24.0.0",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
pythonpath = ["."]

[tool.ruff]
line-length = 100